jesd: jesd bringup fixes on DFE
Multiple fixes for problems found during bringup of JESD
on DFE. This patch will be squashed with base driver patch.

Also added serdes intialization code in jesd driver itself
for bringup, serdes intialization will be moved to serdes
driver. The serdes init part of this driver will not affect
other drivers because  this patch only intializes SERDES1
which is used for JESD only.

Signed-off-by: Pankaj Chauhan <pankaj.chauhan@freescale.com>