jesd [XXX]: Hack for Transmitting just after sync
In order to get TX data flowing out of JESD as soon as it
syncronizes, we program both Tx alignment and Tx AXRF timer
as active low from user space. We don't fire the timers in
driver.

This way the strobes remain active in timer's non fired state,
and JESD starts transmitting out as soon as it syncronizes

Signed-off-by: Pankaj Chauhan <pankaj.chauhan@freescale.com>