tbgen/jesd: Fire TX_ALIGNMENT and AXRF timers
Removed hack for holding back TX_ALIGNMENT and AXRF timers.

Also added one back door for holding back timers if required,
If the TX timer is programmed active low then we do not fire
it in driver.

This way user space can control whether timer needs to be fired
based on timer or to be kept in always ON mode.

Signed-off-by: Pankaj Chauhan <pankaj.chauhan@freescale.com>
Change-Id: I6980d52f32f8c682dce2a69e2dfd5aad671b82fd
Reviewed-on: http://git.am.freescale.net:8181/6589
Reviewed-by: Pankaj Chauhan <Pankaj.Chauhan@freescale.com>
Tested-by: Pankaj Chauhan <Pankaj.Chauhan@freescale.com>