Simulated reset feature added
Updated code per review comments:
- White space removal and line restore
- mach-d4400 Makefile modified to use CONFIG_SOC_D4400
  for restart.c and simreset.c

Updated code per review comments:
- simreset.c/restart.c moved up one dir level.
- CONFIG_SIM_RESET_D4400 references deleted
- CONFIG_SOC_D4400 flag is used to qualify compile of simreset.c and restart.c

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>

The simulated reset feature is added to allow a reset of the cpu without
reseting the module registers.  This feature allows cpri daisychain to
continue working throughout the reset and bootup process.  By default a
reboot command in Linux uses the watchdog timeout to send a hard reset
signal to the cpu.  An ioctl call has been added to the Qixis driver to
allow the user to change the reboot to use simulated reset.

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: I71b3b0fda233aa319c71634de4004be1c1fe89c9
Reviewed-on: http://git.am.freescale.net:8181/30996
Reviewed-by: Nicholas J Spence <Nick.Spence@freescale.com>
Tested-by: Nicholas J Spence <Nick.Spence@freescale.com>