Sfp tx_disable set according to gpio direction
Per review comments, the tx_disable gpio is set as an output high
only if the gpio direction is detected to be an input.  If it is
already an output, then nothing is done to the gpio.

The sfp driver was modified so that tx_disable gpio signal is set
according to the reboot method.  If reboot is cpu reset OR tx_disable
gpio is an input, then set tx_disable to an output high (which disables
the sfp module).  Else if the reboot is a simulated reset, then don't do
anything to the tx_disable gpio.  This change is needed to allow cpri
diasychain to work throught the bootup process.

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: I33ee0c70ed4a578180147a8c8898c4be5af239ed
Reviewed-on: http://git.am.freescale.net:8181/30999
Reviewed-by: Nicholas J Spence <Nick.Spence@freescale.com>
Tested-by: Nicholas J Spence <Nick.Spence@freescale.com>