AFD4400 rev1.1: remove old BGR enable writes to GCR72
GCR Rev 1.1 GCR72 controlled the BGR enables. These bits were re-
purposed in rev 1.1. GCR72 is already set correctly in u-boot based on
the silicon version so all of these writes are redundant.

Rather than making the gcr code check the silicon version it is cleaner
to remove the code and rely on u-boot to set GCR72 correctly.

Signed-off-by: Nick Spence <rxtd10@freescale.com>
Change-Id: Ie19212f3581a746b2cb327c70932b2b774fba59b
Reviewed-on: http://git.am.freescale.net:8181/35194
Reviewed-by: Nicholas J Spence <Nick.Spence@freescale.com>
Tested-by: Nicholas J Spence <Nick.Spence@freescale.com>