Updates to xcvr detect for EVB revA, 8/16-bit addr eeprom read
Modifications to xcvr card detect for EVB revA:
- If two fmc connectors are occupied then the xcvr driver detect is
skipped.
- If a xcvr node specified that it uses two fmc connectors, then detect
for that driver is skipped.

These rules stem from the fact that revA does not support wideband and
if a jesd loopback card is inserted we don't want to load any drivers
for it.  For all other boards where eeprom is powered, IPMI info is used
for xcvr detection.

Updates to eeprom access:
- Can handle eeprom > 256 bytes in size and up to 256KB.  Jesd loopback
eeprom is 16KB and requires 2-byte addressing.
- Separated code that reads/populate eeprom info from dts node.
- Separated code that reads from eeprom.

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: Ic3eb5370003748ff2bd6efa9612165583663e938
Reviewed-on: http://git.am.freescale.net:8181/35332
Reviewed-by: Nicholas J Spence <Nick.Spence@freescale.com>
Tested-by: Nicholas J Spence <Nick.Spence@freescale.com>