Updated wideband eeprom size in dts
With eeprom upgraded on the wideband board from 256 bytes to 64KB
(1Mbit), the dts must be updated to the new eeprom size so that the
driver can correctly access it with 2 byte address instead of 1.  This
update has been tested with the 64KB eeprom on the wideband board with
RDB board.

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: I9bf081dc687a0e5a58ee9715dfba109c3f2f73ee
Reviewed-on: http://git.am.freescale.net:8181/35929
Reviewed-by: David Henning <Dave.Henning@freescale.com>
Tested-by: David Henning <Dave.Henning@freescale.com>