Device tree compiler tool update, v1.4 with overlay option
The device tree compiler tool has been updated to v1.4.  Also, the
overlay feature has been added.  This feature is enabled with the dtc
tool option "-@" which adds the symbolic (human readable) node
information in the .dtb file.

The "-@" option was added in scripts/Makefile.lib so that the
d4400-evb.dtb file has the symbolic node info.  The dtb file is about
20-25% larger due to the symbolic node info added.

Besides the overlay option, v1.4 dtb output is compatible with v1.2
which was the old version.  The version can be displayed as follows:

$~/dfe/linux/scripts/dtc -v
Version: DTC 1.4.0-overlay-f8cb5dd94

Note that DTC v1.4 source code from public git repository does not have
the overlay feature.  The overlay feature is a patch that has to be
applied to v1.4.  Hence, the text "overlay" was added to the version to
differentiate it from the "regular" v1.4.

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: Idbe652e00c11216b355023155dc59014ff7e54e2
Reviewed-on: http://git.am.freescale.net:8181/39871
Reviewed-by: David Henning <Dave.Henning@freescale.com>
Tested-by: David Henning <Dave.Henning@freescale.com>