Board init to continue in event of bad IPMI data
Changes to d4400 system driver:

- For 4T4R POC board, board initialization is allowed to continued in
the event of bad or missing IPMI data in system eeprom.

- Added mutex lock to sysfs callback functions.

...and other misc changes.

A set of functions were written and tested to give ability to program
IPMI data into system via sysfs.  This code is not being used for now so
they are placed at the end of the fsl-d4400-sys.c file for archive.

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: I4217a47f0fab44ed6ec9965e9385479a1b0274e1
Reviewed-on: http://git.am.freescale.net:8181/40030
Reviewed-by: David Henning <Dave.Henning@freescale.com>
Tested-by: David Henning <Dave.Henning@freescale.com>