Jcpll spi-gpio signal overlay for POC boards
Device tree overlay for jcpll spi-gpio signals are added to the D4400
system driver so that the spi miso/mosi can be set to normal or
reversed depending on which board is running.  For 4t4r revA, the
overlay is applied to swap miso/mosi signals.  For all other boards,
the overlay is applied for normal miso/mosi signal configuration. Note
that proper IPMI data needs to be present on the board's eeprom.

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: Iff749483aada355a75ca059c11bf4365d9e738db
Reviewed-on: http://git.am.freescale.net:8181/40396
Reviewed-by: David Henning <Dave.Henning@freescale.com>
Tested-by: David Henning <Dave.Henning@freescale.com>