KERNEL: WB2 added LE pulse for lmx2492 and pe43711 devices
The LMX2492 and PE43711 devices require that the host generate an extra
pulse on the slave select or latch-enable (LE) signal in order to latch
the new value and make it effective.

This patch adds that feature to generate the LE pulse after the SPI
write is completed.  The LE pulse is generated for EACH SPI write to a
register.

The implementation performs the extra LE pulse by configuring the SPI
slave select to gpio function.  Then, the LE pulse starts at about ~55
usec after the SPI write, a low of  ~9 usec, high of ~3.5 usec, and a
low of ~2.5 usec.  At this point LE goes high as we return the signal
back to SPI slave select functionality.

Testing was perform to verify that the code can pulse the gpio signal of
each of these devices:

   SPI6 SS0: lmx2492 DCC/RX LO, U78, LE pin 16 -> R484  VERIFIED
   SPI6 SS1: lmx2492 TX LO, U89, LE pin 16 -> R546	VERIFIED
   SPI6 SS2: lmx2492 SRX/RX LO, U103, LE pin 16 -> R615 VERIFIED

NOTE PE43711: The R values listed here may not match board value.
   Look for components to probe for on pin 16 (6 pins per side).

   SPI7 SS0: pe43711 TX1, U24, LE pin 16 -> R184 VERIFIED
   SPI8 SS0: pe43711 TX2, U28, LE pin 16 -> R209 VERIFIED
   SPI5 SS0: pe43711 SRX, U46, LE pin 16 -> R289 VERIFIED

Signed-off-by:  Hung Tran <hung.h.tran@freescale.com>
Change-Id: I34e597613b1676285bf68e7c93a2fee0a0c9c4ff
Reviewed-on: http://git.am.freescale.net:8181/41273
Reviewed-by: Dave Henning <Dave.Henning@freescale.com>
Tested-by: Dave Henning <Dave.Henning@freescale.com>