CPRI: Add support for non-contiguous AxC samples
In the past, only mapping entry was allowed for
each AxCs.  However, the hardware supports multiple
entries.

This patch allows multiple entries per AxC.  This makes
it possible for samples to be interleaved within the
basic frame in any pattern.

The changes to the code are not too significant.  Really,
all that has changed is the error checking code.  Specifically:

 1. Two or more entries can have the same ID
 2a. Entries with the same ID must have the same
    buffer address
 2b. Entires with the same buffer address must
   have the same ID
 3. The check for >12 AxCs in a MEMBLK must take
    into account duplicate IDs.